Added Xcode Project; Created a NavigationViewController in Storyboard
Removed the existing ViewController storyboard and class that were created with the project. Created a new UINavigationController and 2 UIViewControllers that will set up the foundation of this project. The UINavigationController is responsible for pushing and popping the ColorsDetailVC onto the stack via the ColorsTableVC. Setup the respective VC classes that the ColorsDetailVC and ColorsTableVC are linked to, and setup a basic button for transitioning between the views.